requirements.txt
aiohttp==3.9.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
frozenlist==1.4.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
multidict==6.0.4
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pypdfium2==4.24.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
twilio==8.10.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
yarl==1.9.3
